+++ themes/ideaspace-3d-model/views/index.blade.php
@@ -25,17 +25,22 @@
     <meta property="og:url" content="{{ \Request::url() }}" />
 
     <link rel="stylesheet" href="{{ url($theme_dir . '/css/style.css') }}">
-    <script src="{{ url($theme_dir . '/js/aframe/aframe-v0.7.1.min.js') }}"></script>
+    <script src="{{ url($theme_dir . '/js/aframe/aframe-v0.8.2.min.js') }}"></script>
+    <script src="{{ url($theme_dir . '/js/aframe-teleport-controls/aframe-teleport-controls.min.js') }}"></script>
     <script src="{{ url($theme_dir . '/js/gradientsky.min.js') }}"></script>
 		<script src="{{ url($theme_dir . '/js/aframe-html-shader-component/aframe-html-shader.min.js') }}"></script>
     <script src="{{ url($theme_dir . '/js/aframe-extras/aframe-extras.loaders.min.js') }}"></script>
-    <script src="{{ url($theme_dir . '/js/aframe-orbit-controls/aframe-orbit-controls-component.js') }}"></script>
+    <script src="{{ url($theme_dir . '/js/aframe-orbit-controls/aframe-orbit-controls-component.min.js') }}"></script>
     <script src="{{ url($theme_dir . '/js/kframe/aframe-look-at-component.min.js') }}"></script>
     <script src="{{ url($theme_dir . '/js/isvr-vr-mode.js') }}"></script>
     <script src="{{ url($theme_dir . '/js/isvr-model-center.js') }}"></script>
-    <script src="{{ url($theme_dir . '/js/isvr-hotspot.js') }}"></script>
-    <script src="{{ url($theme_dir . '/js/isvr-annotation.js') }}"></script>
-    <script src="{{ url($theme_dir . '/js/isvr-teleportation.js') }}"></script>
+
+<script src="{{ url($theme_dir . '/js/kframe/aframe-log-component.min.js') }}"></script>
+<script>
+window.addEventListener('vrdisplayactivate', function() {
+AFRAME.log('VR Display Activate');
+});
+</script>
 </head>
 <body>